herbivore and predator behavior

+++ src/main/java/org/example/entity/creature/Herbivore.java
@@ -7,7 +7,7 @@ public class Herbivore extends Creature {
     private static final int DEFAULT_SPEED = 2;
     private static final int DEFAULT_HEALTH = 100;
 
-    public Herbivore(int speed, double health) {
+    public Herbivore(int speed, int health) {
         super(speed, health, new HerbivoreBehavior());
     }
